% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel.R
\name{all_available_panels}
\alias{all_available_panels}
\title{Get all available gene panels in a database}
\usage{
all_available_panels(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query}
}
\value{
a dataframe of metadata regarding each available panel
}
\description{
Get all available gene panels in a database
}
