% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_genes_lookup}
\alias{all_genes_lookup}
\title{A table to lookup Hugo Symbol by Entrez ID}
\format{
A data frame
\describe{
\item{entrezGeneId}{entrez Gene Id}
\item{hugoGeneSymbol}{Gene Hugo Symbol}
\item{type}{Gene type}
}
}
\usage{
all_genes_lookup
}
\description{
A table to lookup Hugo Symbol by Entrez ID
}
\keyword{datasets}
