% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_studies.R
\name{available_sample_lists}
\alias{available_sample_lists}
\title{Get All Sample Lists Available For a Study}
\usage{
available_sample_lists(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{A character string indicating which study ID should be searched.
Only 1 study ID allowed.}

\item{base_url}{The database URL to query
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of patient_ids in a given study
}
\description{
Get All Sample Lists Available For a Study
}
\examples{
\dontrun{
set_cbioportal_db("public")
available_sample_lists(study_id = "acc_tcga")
}

}
