% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_database.R
\name{available_studies}
\alias{available_studies}
\title{Get Metadata on All Available Studies in a Database}
\usage{
available_studies(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of available studies and their metadata
}
\description{
Get Metadata on All Available Studies in a Database
}
\examples{
\dontshow{if (!httr::http_error("www.cbioportal.org/api")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_cbioportal_db("public")
available_studies()
\dontshow{\}) # examplesIf}
}
