% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportalr-package.R
\docType{package}
\name{cbioportalR-package}
\alias{cbioportalR}
\alias{cbioportalR-package}
\title{cbioportalR: Browse and Query Clinical and Genomic Data from cBioPortal}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides R users with direct access to genomic and clinical data from the 'cBioPortal' web resource via user-friendly functions that wrap 'cBioPortal's' existing API endpoints \url{https://www.cbioportal.org/api/swagger-ui/index.html}. Users can browse and query genomic data on mutations, copy number alterations and fusions, as well as data on tumor mutational burden ('TMB'), microsatellite instability status ('MSI'), 'FACETS' and select clinical data points (depending on the study). See \url{https://www.cbioportal.org/} and Gao et al., (2013) \doi{10.1126/scisignal.2004088} for more information on the cBioPortal web resource.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/karissawhiting/cbioportalR}
  \item \url{https://www.karissawhiting.com/cbioportalR/}
  \item Report bugs at \url{https://github.com/karissawhiting/cbioportalR/issues}
}

}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

}
\keyword{internal}
