% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{check_for_saved_db}
\alias{check_for_saved_db}
\title{Check for cBioPortal DB}
\usage{
check_for_saved_db()
}
\value{
a saved data base URLir NULL if none exists
}
\description{
This function  checks for and retrieves cBioPortal URL System Environment variable "CBIOPORTAL_URL"
}
\examples{
check_for_saved_db()

}
