% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{.get_data_by_sample}
\alias{.get_data_by_sample}
\title{Get Mutations By Sample ID}
\usage{
.get_data_by_sample(
  sample_id = NULL,
  study_id = NULL,
  molecular_profile_id = NULL,
  sample_study_pairs = NULL,
  data_type = c("mutations", "cna"),
  genes = NULL,
  base_url = NULL
)
}
\arguments{
\item{sample_id}{a vector of sample IDs (character)}

\item{study_id}{A string indicating the study ID from which to pull data. If no study ID, will
guess the study ID based on your URL and inform. Only 1 study ID can be passed. If mutations from
more than 1 study needed, see \code{sample_study_pairs}}

\item{molecular_profile_id}{A string indicating the molecular profile ID from which to pull data. If ID supplied, will
guess the molecular profile ID based on the study ID. Only 1 molecular profile ID can be passed. If mutations from
more than 1 study needed, see \code{sample_study_pairs}}

\item{sample_study_pairs}{A dataframe with columns: \code{sample_id}, \code{study_id} and \code{molecular_profile_id} (optional).
This can be used in place of \code{sample_id}, \code{study_id}, \code{molecular_profile_id} arguments above if you
need to pull samples from several different studies at once.}

\item{genes}{A vector of entrez ids. If NULL, will return results for all
IMPACT genes (see \code{gnomeR::impact_gene_info})}

\item{base_url}{The database URL to query
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{

}
\description{
Get Mutations By Sample ID
}
\examples{
.get_data_by_sample(sample_id = c("TCGA-OR-A5J2-01","TCGA-OR-A5J6-01" ),
 study_id = "acc_tcga")

df_pairs <- data.frame(
"sample_id" = c("s_C_36924L_P001_d",
"s_C_03LNU8_P001_d"),
"study_id" = c("prad_msk_2019"))

.get_data_by_sample(sample_study_pairs = df_pairs)
.get_data_by_sample(sample_study_pairs = df_pairs, genes = 7157)
.get_data_by_sample(sample_study_pairs = df_pairs, data_type = "cna")

}
