% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_alias}
\alias{get_alias}
\title{Get Gene Name Alias for a Given Hugo Symbol}
\usage{
get_alias(hugo_symbol, ...)
}
\arguments{
\item{hugo_symbol}{a hugo symbol for which to return aliases}

\item{...}{Not used}
}
\value{
A character string with all aliases
}
\description{
This only works 1 gene at a time right now
}
\examples{
get_cbioportal_db('public')
get_alias("FGFR3")

}
