% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_cbioportal_db}
\alias{get_cbioportal_db}
\title{Connect to cBioPortal DB}
\usage{
get_cbioportal_db(db = NULL)
}
\arguments{
\item{db}{The database URL to use as base URL for calls, or "public" for https://www.cbioportal.org/}
}
\description{
This function sets a base cBioPortal url
}
\examples{
\dontrun{
get_cbioportal_db(db = "public")
}

}
