% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_cbioportal_url}
\alias{get_cbioportal_url}
\title{Get Database URL to Use in Functions}
\usage{
get_cbioportal_url()
}
\value{

}
\description{
Pulls the set URL from the internal package environement
}
