% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_attributes}
\alias{get_clinical_attributes}
\title{Get all clinical attributes available for a study}
\usage{
get_clinical_attributes(study_id = NULL)
}
\arguments{
\item{study_id}{cbioportal study ID}
}
\value{
a data frame of available clinical attributes for that study
}
\description{
Get all clinical attributes available for a study
}
\examples{
get_cbioportal_db('public')
get_clinical_attributes("acc_tcga")
}
