% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_by_patient}
\alias{get_clinical_by_patient}
\title{Get clinical data by attribute study ID and sample ID}
\usage{
get_clinical_by_patient(
  study_id = NULL,
  sample_id = NULL,
  clinical_attribute = "CANCER_TYPE",
  base_url = NULL
)
}
\arguments{
\item{study_id}{study ID}

\item{sample_id}{a single tumor sample ID}

\item{clinical_attribute}{a specific clinical attribute}

\item{base_url}{The database URL to query}
}
\value{
a dataframe of a specific clinical attribute
}
\description{
Get clinical data by attribute study ID and sample ID
}
\examples{
get_clinical_by_patient(study_id = "acc_tcga", sample_id = "TCGA-OR-A5J2-01",
 clinical_attribute = "CANCER_TYPE", base_url = 'www.cbioportal.org/api')
}
