% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_by_study}
\alias{get_clinical_by_study}
\title{Get all available clinical data for a specified study}
\usage{
get_clinical_by_study(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{study ID}
}
\value{
a dataframe of all available clinical attributes and their values
}
\description{
Get all available clinical data for a specified study
}
\examples{
get_clinical_by_study(study_id = "acc_tcga", base_url = 'www.cbioportal.org/api')

}
