% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cna.R
\name{get_cna}
\alias{get_cna}
\title{Get copy number alterations by sample ID or study ID}
\usage{
get_cna(study_id = NULL, sample_ids = NULL, panel = NULL, genes = "all_impact")
}
\arguments{
\item{study_id}{A character vector of lenth 1 indicating study_id. See \code{get_studies()} to see available studies.}

\item{sample_ids}{A character vector of sample ids}

\item{panel}{OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL,
the panel will be looked up with \code{get_panel()} and only genes in that panel will be returned.}

\item{genes}{A list of genes to query. default is all impact genes.}
}
\value{
A dataframe of CNA for each sample ID (in maf file format)
}
\description{
Get copy number alterations by sample ID or study ID
}
\examples{
\dontrun{
get_cbioportal_db('public')
get_cna(sample_ids = c("TCGA-19-5959-01", "TCGA-18-4083-01"))
}

}
