% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_gene_id}
\alias{get_gene_id}
\title{Get Entrez Gene ID for a given set of Hugo Symbols}
\usage{
get_gene_id(hugo_symbol = NULL, base_url = NULL)
}
\arguments{
\item{hugo_symbol}{a hugo symbol for which to return an Entrez Gene ID}

\item{base_url}{The database URL to query}
}
\value{
A dataframe with Entrez Gene IDs and Hugo Symbols
}
\description{
This only works 1 gene at a time right now
}
\examples{

get_gene_id("FGFR3", base_url = 'www.cbioportal.org/api')

}
