% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{get_gene_panel}
\alias{get_gene_panel}
\title{Retrieve Genes Included For a Specified Panel ID}
\usage{
get_gene_panel(panel_id = NULL, base_url = NULL)
}
\arguments{
\item{panel_id}{name of panel. See \code{available_gene_panels()} to get panel ID}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of genes in a specified panel
}
\description{
Retrieve Genes Included For a Specified Panel ID
}
