% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Get A List of Genes for a Specified Database}
\usage{
get_genes(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query}
}
\value{
A dataframe of gene ids, hugo symbols, and gene types
}
\description{
Get A List of Genes for a Specified Database
}
\examples{
\dontrun{
get_genes(base_url = 'www.cbioportal.org/api')
}
}
