% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-genetics.R
\name{get_genetics}
\alias{get_genetics}
\title{Get all genetic data for a study or list of IDs}
\usage{
get_genetics(
  sample_ids = NULL,
  study_id = NULL,
  genes = NULL,
  panel = NULL,
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE,
  seg = FALSE
)
}
\arguments{
\item{sample_ids}{A character vector of sample ids}

\item{study_id}{A character vector of length 1 specifying study to query. See \code{get_studies()} to see list of available studies.}

\item{genes}{A list of genes to query. default is all impact genes.}

\item{panel}{OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL,
the panel will be looked up with \code{get_panel()} and only genes in that panel will be returned.}

\item{mutations}{Boolean specifying if mutation data should be fetched. Default is TRUE.}

\item{fusions}{Boolean specifying if fusion data should be fetched. Default is TRUE.}

\item{cna}{Boolean specifying if cna data should be fetched. Default is TRUE.}

\item{seg}{Boolean specifying if segmentation data should be fectched. Default is FALSE.}
}
\value{
A list of dataframes containing genetic data.
}
\description{
Get all genetic data for a study or list of IDs
}
\examples{
\dontrun{
# IMPACT #
get_genetics(sample_ids = c("P-0000004-T01-IM3", "P-0000012-T02-IM3"),
genes = 207)

# TCGA #
get_genetics(sample_ids =  c("TCGA-17-Z023-01","TCGA-02-0003-01","TCGA-02-0055-01"),
 database = "tcga")
}
}
