% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_genes_or_panel.R
\name{get_hugo_symbol}
\alias{get_hugo_symbol}
\title{Get Hugo Symbol for a given set of Entrez IDs}
\usage{
get_hugo_symbol(entrez_id = NULL, base_url = NULL)
}
\arguments{
\item{entrez_id}{a character or numeric vector of Entrez gene IDs}

\item{base_url}{The database URL to query}
}
\value{
A dataframe with Entrez Gene IDs and Hugo Symbols
}
\description{
Get Hugo Symbol for a given set of Entrez IDs
}
\examples{
\dontshow{if (!httr::http_error("www.cbioportal.org/api")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

get_hugo_symbol(entrez_id = 2261, base_url = 'www.cbioportal.org/api')
get_hugo_symbol(entrez_id = c(2261, 7157) , base_url = 'www.cbioportal.org/api')
\dontshow{\}) # examplesIf}
}
