% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_cna_by_study.R, R/mutation.R,
%   R/mutation_and_cna.R
\name{get_mutation_by_study}
\alias{get_mutation_by_study}
\title{Get Mutations By Study ID}
\usage{
get_mutation_by_study(
  study_id = NULL,
  molecular_profile_id = NULL,
  base_url = NULL
)

get_mutation_by_study(
  study_id = NULL,
  molecular_profile_id = NULL,
  base_url = NULL
)

get_mutation_by_study(
  study_id = NULL,
  molecular_profile_id = NULL,
  base_url = NULL
)
}
\arguments{
\item{study_id}{A study ID to query mutations. If NULL, guesses study ID based
on molecular_profile_id.}

\item{molecular_profile_id}{a molecular profile to query mutations.
If NULL, guesses molecular_profile_id based on study ID.}

\item{base_url}{The database URL to query
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{





}
\description{
Get Mutations By Study ID

Get Mutations By Study ID

Get Mutations By Study ID
}
\examples{
get_mutation_by_study(study_id = "prad_msk_2019")
get_mutation_by_study(molecular_profile_id = "prad_msk_2019_mutations")


get_mutation_by_study(study_id = "prad_msk_2019")
get_mutation_by_study(molecular_profile_id = "prad_msk_2019_mutations")


get_mutation_by_study(study_id = "prad_msk_2019")
get_mutation_by_study(molecular_profile_id = "prad_msk_2019_mutations")


}
