% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutations.R
\name{get_mutations}
\alias{get_mutations}
\title{Get gene mutations by sample ID or Study}
\usage{
get_mutations(
  sample_id = NULL,
  study_id = NULL,
  panel = NULL,
  genes = NULL,
  ...
)
}
\arguments{
\item{sample_id}{A character vector of sample ids. Can be NULL if a \code{study_id} is passed. If sample IDs are
passed with no \code{study_id}, IMPACT studies will be queried by default.
If non IMPACT ID samples passed, a \code{study_id} is required as well to retrieve data.}

\item{study_id}{A character vector study IDs. See \code{get_studies()} to see available studies. If passed with no specified \code{sample_id}
all samples for specified studies will be returned.}

\item{panel}{OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL,
the panel will be looked up with \code{get_panel()} and only genes in that panel will be returned.}

\item{genes}{A list of genes to query. default is all impact genes.}

\item{...}{Not used}
}
\value{
A dataframe of mutations for each sample ID (in maf file format)
}
\description{
Get gene mutations by sample ID or Study
}
\examples{
get_cbioportal_db("public")
get_mutations(sample_id = c("P-0005217-T03-IM5", "P-0038798-T01-IM6"))

}
