% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel.R
\name{get_panel}
\alias{get_panel}
\title{Retrieve data for a specific panel}
\usage{
get_panel(panel_id)
}
\arguments{
\item{panel_id}{name of panel}
}
\value{
A dataframe of genes in a specified panel
}
\description{
Retrieve data for a specific panel
}
