% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_samples.R
\name{get_panel_by_sample}
\alias{get_panel_by_sample}
\title{Get Gene Panel by study ID and sample ID}
\usage{
get_panel_by_sample(study_id = NULL, sample_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{study ID}

\item{sample_id}{a vector of sample IDs}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
a dataframe of a specific clinical attribute
}
\description{
Get Gene Panel by study ID and sample ID
}
\examples{
get_panel_by_sample(study_id = "blca_plasmacytoid_mskcc_2016",
 sample_id = "DS-sig-010-P2",
 base_url = 'www.cbioportal.org/api')

}
