% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_id.R, R/get_sample_ids.R
\name{get_sample_ids}
\alias{get_sample_ids}
\title{Get sample IDs for a given set of patient IDs}
\usage{
get_sample_ids(patient_ids = NULL, study_id = NULL)

get_sample_ids(patient_ids = NULL, study_id = NULL)
}
\arguments{
\item{patient_ids}{A character string of sample IDs to query}

\item{study_id}{A character string of study_id indicating which study to
look for samples in. Is NULL by default and looks in msk_impact_2017 study if you
do no specify a study and are using public cbioportal. If you are connected to MSK
cbioportal it will look in the most up to date IMPACT study (which is not public).}
}
\value{
A dataframe of patient IDs and corresponding sample IDs. If patient
has multiple samples, there will be multiple rows per patient.

A dataframe of patient IDs and corresponding sample IDs. If patient
has multiple samples, there will be multiple rows per patient.
}
\description{
Get sample IDs for a given set of patient IDs

Get sample IDs for a given set of patient IDs
}
\examples{
\dontrun{
get_cbioportal_db("public")
get_sample_ids(patient_ids = c("P-0000034", "P-0000036"))
}

\dontrun{
get_cbioportal_db("public")
get_sample_ids(patient_ids = c("P-0000034", "P-0000036"))
}

}
