% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments.R
\name{get_segments}
\alias{get_segments}
\title{Get segmentation data}
\usage{
get_segments(sample_ids = NULL, study_id = "all_tcga_studies")
}
\arguments{
\item{sample_ids}{A character vector of sample ids.}

\item{study_id}{A character vector of lenght 1 indicating study_id.
Default is "all_tcga_studies" which queries all TCGA cancer studies, unless you
are connected to MSKCC host in which case default is "mskimpact".}
}
\value{
Returns a dataframe of segmentation data
}
\description{
Get segmentation data
}
\examples{
get_cbioportal_db('public')
get_segments(sample_ids = "TCGA-OR-A5J2-01", study_id = "acc_tcga")

}
