% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_studies.R
\name{get_studies}
\alias{get_studies}
\title{Get information about available studies in a cBioPortal instance}
\usage{
get_studies(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{Character vector of one single study ID (e.g. \code{"mskimpact"}) or NULL.
Default is NULL which returns all available studies. Currently multiple study ids are not supported.}
}
\value{
A tibble of study meta data
}
\description{
Get information about available studies in a cBioPortal instance
}
\examples{
get_cbioportal_db("public")
get_studies()
get_studies(study_id = "msk_impact_2017")

}
