% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{get_study_info}
\alias{get_study_info}
\title{Get Metadata on All Available Studies in Database or a Specified Study}
\usage{
get_study_info(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{one or more study IDs (seee \code{available_studies()} to lookup IDs)}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of study metadata
}
\description{
Get Metadata on All Available Studies in Database or a Specified Study
}
