% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impact_gene_info}
\alias{impact_gene_info}
\title{IMPACT Gene Meta Data}
\format{
A data frame with 470 genes
\describe{
\item{hugo_symbol}{Factor w/ 574 levels,
Column containing all HUGO symbols genes included in IMPACT}
\item{entrez_id}{Integer, contains all Entrez IDs for genes included in IMPACT}
\item{platform_341}{Character, indicates whether each gene was included in
IMPACT platform 341. Options are \code{included} and \verb{not included}}
\item{platform_410}{Character, indicates whether each gene was included in
IMPACT platform 410. Options are \code{included} and \verb{not included}}
\item{platform_468}{Character, indicates whether each gene was included in
IMPACT platform 468. Options are \code{included} and \verb{not included}}
\item{alias}{A nested dataframe of aliases for each gene and corresponding entrez gene ids for aliases if they exist}
}
}
\source{
\url{https://cbioportal.org/}
}
\usage{
impact_gene_info
}
\description{
Dataframe labeling all genes found in IMPACT along with their corresponding
platform and Entrez ID.
}
\keyword{datasets}
