% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_panel.R
\name{all_available_panels}
\alias{all_available_panels}
\title{Get all available gene panels in a database}
\usage{
all_available_panels()
}
\value{
a dataframe of metadata regarding each available panel
}
\description{
Get all available gene panels in a database
}
