% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{available_clinical_attributes}
\alias{available_clinical_attributes}
\title{Get all available clinical attribute IDs for a study}
\usage{
available_clinical_attributes(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{cbioportal study ID}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
a data frame of available clinical attributes for that study
}
\description{
Get all available clinical attribute IDs for a study
}
\examples{
available_clinical_attributes("acc_tcga", base_url = 'www.cbioportal.org/api')

}
