% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\name{available_gene_panels}
\alias{available_gene_panels}
\title{Get Available Gene Panels For a Database}
\usage{
available_gene_panels(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
a dataframe of metadata regarding each available panel
}
\description{
Get Available Gene Panels For a Database
}
