% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_profiles.R
\name{available_profiles}
\alias{available_profiles}
\title{See what genetic data is available for a given study}
\usage{
available_profiles(study_id = NULL)
}
\arguments{
\item{study_id}{A character vector of lenth 1 indicating study_id. See \code{get_studies()} to see available studies.}
}
\value{
A dataframe of available genetic profiles and their names
}
\description{
See what genetic data is available for a given study
}
\examples{
get_cbioportal_db('public')
available_profiles("acc_tcga")

}
