% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_studies.R
\name{available_samples}
\alias{available_samples}
\title{Get All Sample IDs in a Study}
\usage{
available_samples(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{A character string indicating which study ID should be searched.
Only 1 study allowed. If NULL, we will guess a default study ID based on your database URL.}

\item{base_url}{The database URL to query
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of sample_ids in a given study
}
\description{
Get All Sample IDs in a Study
}
\examples{
\dontrun{
set_cbioportal_db("public")
available_samples(study_id = "acc_tcga")
}

}
