% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportalr-package.R
\docType{package}
\name{cbioportalR-package}
\alias{cbioportalR}
\alias{cbioportalR-package}
\title{cbioportalR: An R package to access data on cBioPortal}
\description{
allows you to access cBioPortal's genomic and clinical data directly through R.
    The package wraps cBioPortal's existing API endpoints in R so R users can easily leverage cBioPortal's API.
    Using these functions, you can access genomic data on mutations, copy number alterations and fusions as well as data on tumor mutational burden (TMB),
    microsatellite instability status (MSI) and select clinical data points (depending on the study).
}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

Authors:
\itemize{
  \item Axel Martin
  \item Arshi Arora
}

}
\keyword{internal}
