% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbioportalr-package.R
\docType{package}
\name{cbioportalR-package}
\alias{cbioportalR}
\alias{cbioportalR-package}
\title{cbioportalR: An R Package to Access Data on cBioPortal}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R client for access to genomic and clinical data from `cBioPortal`. The package wraps 'cBioPortal's' existing API endpoints so R users can easily leverage the existing API. Using these functions, users can access genomic data on mutations, copy number alterations and fusions, as well as data on tumor mutational burden ('TMB'), microsatellite instability status ('MSI'), 'FACETS' data and select clinical data points (depending on the study).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/karissawhiting/cbioportalR}
  \item \url{https://www.karissawhiting.com/cbioportalR/}
  \item Report bugs at \url{https://github.com/karissawhiting/cbioportalR/issues}
}

}
\author{
\strong{Maintainer}: Karissa Whiting \email{karissa.whiting@gmail.com} (\href{https://orcid.org/0000-0002-4683-1868}{ORCID})

}
\keyword{internal}
