% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_cbioportal_token}
\alias{get_cbioportal_token}
\title{Get cBioPortal Access Token}
\usage{
get_cbioportal_token()
}
\description{
This function retrieves cBioPortal token System Environment variable "CBIOPORTAL_TOKEN"
}
\examples{
\dontrun{
get_cbioportal_token()
}

}
