% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_cbioportal_token}
\alias{get_cbioportal_token}
\title{Get cBioPortal Access Token}
\usage{
get_cbioportal_token()
}
\description{
Convenience function that retrieves cBioPortal token System Environment variable "CBIOPORTAL_TOKEN"
}
\examples{
\dontshow{if (!httr::http_error("www.cbioportal.org/api")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_cbioportal_token()
\dontshow{\}) # examplesIf}
}
