% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{get_cbioportal_token}
\alias{get_cbioportal_token}
\title{Get cBioPortal Access Token}
\usage{
get_cbioportal_token()
}
\value{
Returns a string with cBioPortal token if successfully authenticated, or
a warning that token could not be found.
}
\description{
Convenience function that retrieves cBioPortal token System Environment variable "CBIOPORTAL_TOKEN"
}
\examples{
\dontrun{
get_cbioportal_token()
}

}
\author{
Karissa Whiting, Daniel D. Sjoberg
}
