% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clinical_data.R
\name{get_clinical_by_sample}
\alias{get_clinical_by_sample}
\title{Get clinical data by attribute, study ID and sample ID}
\usage{
get_clinical_by_sample(
  study_id = NULL,
  sample_id = NULL,
  clinical_attribute = NULL,
  base_url = NULL
)
}
\arguments{
\item{study_id}{study ID}

\item{sample_id}{a vector of sample IDs}

\item{clinical_attribute}{one or more clinical attributes for your study.
If none provided, will return all attributes available for
that study (\verb{available_clinical_attributes(<study_id>)})}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
a dataframe of a specific clinical attribute
}
\description{
Get clinical data by attribute, study ID and sample ID
}
\examples{
get_clinical_by_sample(study_id = "acc_tcga", sample_id = "TCGA-OR-A5J2-01",
 clinical_attribute = "CANCER_TYPE", base_url = 'www.cbioportal.org/api')

}
