% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studies.R
\name{get_clinical_by_study}
\alias{get_clinical_by_study}
\title{Get all available clinical data for a specified study}
\usage{
get_clinical_by_study(study_id = NULL, base_url = NULL)
}
\arguments{
\item{study_id}{study ID}

\item{base_url}{The database URL to query.
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
a dataframe of all available clinical attributes and their values
}
\description{
Get all available clinical data for a specified study
}
\examples{
get_clinical_by_study(study_id = "acc_tcga", base_url = 'www.cbioportal.org/api')

}
