% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Get a list of all available genes}
\usage{
get_genes(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query}
}
\value{
A dataframe of gene ids, hugo dymbols, and gene types
}
\description{
Get a list of all available genes
}
\examples{
\dontrun{
get_genes(base_url = 'www.cbioportal.org/api')
}
}
