% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_database.R
\name{get_genes}
\alias{get_genes}
\title{Get A List of Genes for a Specified Database}
\usage{
get_genes(base_url = NULL)
}
\arguments{
\item{base_url}{The database URL to query
If \code{NULL} will default to URL set with \verb{set_cbioportal_db(<your_db>)}}
}
\value{
A dataframe of gene ids, hugo symbols, and gene types
}
\description{
Get A List of Genes for a Specified Database
}
\examples{
\dontrun{
get_genes(base_url = 'www.cbioportal.org/api')
}
}
