% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{get_genetics}
\alias{get_genetics}
\title{Get all genetic data for a study or list of IDs}
\usage{
get_genetics(
  sample_ids = NULL,
  study_id = NULL,
  genes = NULL,
  panel = NULL,
  mutations = TRUE,
  fusions = TRUE,
  cna = TRUE,
  base_url = NULL
)
}
\arguments{
\item{sample_ids}{A character vector of sample ids. Can be NULL if a \code{study_id} is passed. If sample IDs are
passed with no \code{study_id}, IMPACT studies will be queried by default.
If non IMPACT ID samples passed, a \code{study_id} is required as well to retrieve data.}

\item{study_id}{A character vector study IDs. See \code{get_studies()} to see available studies. If passed with no specified \code{sample_id}
all samples for specified studies will be returned.}

\item{genes}{A vector of Entrez IDs of genes to query. default is all impact genes.}

\item{panel}{OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL,
the panel will be looked up with \code{get_panel()} and only genes in that panel will be returned.}

\item{mutations}{Boolean specifying if mutation data should be fetched. Default is TRUE.}

\item{fusions}{Boolean specifying if fusion data should be fetched. Default is TRUE.}

\item{cna}{Boolean specifying if cna data should be fetched. Default is TRUE.}

\item{base_url}{The database URL to query}
}
\value{
A list of dataframes containing genetic data.
}
\description{
Get all genetic data for a study or list of IDs
}
\examples{
set_cbioportal_db("public")
get_genetics(sample_id = c("P-0005217-T03-IM5", "P-0038798-T01-IM6"))
}
