% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{get_hugo_symbol}
\alias{get_hugo_symbol}
\title{Get Hugo Symbol for a given set of Entrez IDs}
\usage{
get_hugo_symbol(entrez_id = NULL, base_url = NULL)
}
\arguments{
\item{entrez_id}{a character or numeric vector of Entrez gene IDs}

\item{base_url}{The database URL to query}
}
\value{
A dataframe with Entrez Gene IDs and Hugo Symbols
}
\description{
Get Hugo Symbol for a given set of Entrez IDs
}
\examples{

get_hugo_symbol(entrez_id = 2261, base_url = 'www.cbioportal.org/api')
get_hugo_symbol(entrez_id = c(2261, 7157) , base_url = 'www.cbioportal.org/api')
}
