% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mutations.R
\name{get_mutations}
\alias{get_mutations}
\title{Get gene mutations by sample ID or Study}
\usage{
get_mutations(sample_ids = NULL, study_id = NULL, panel = NULL, genes = NULL)
}
\arguments{
\item{sample_ids}{A character vector of sample ids}

\item{study_id}{A character vector of length 1 indicating study_id. See \code{get_studies()} to see available studies.}

\item{panel}{OPTIONAL argument. A character vector of length 1 indicating a specific panel to be used. If not NULL,
the panel will be looked up with \code{get_panel()} and only genes in that panel will be returned.}

\item{genes}{A list of genes to query. default is all impact genes.}
}
\value{
A dataframe of mutations for each sample ID (in maf file format)
}
\description{
Get gene mutations by sample ID or Study
}
\examples{
get_cbioportal_db("public")
get_mutations(sample_ids = c("TCGA-19-5959-01", "TCGA-18-4083-01"))

}
