% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{test_cbioportal_db}
\alias{test_cbioportal_db}
\title{Test the Database Connection Anytime During your R Session}
\usage{
test_cbioportal_db()
}
\description{
Helps troubleshoot API issues during an R session
}
\examples{
set_cbioportal_db("public")
test_cbioportal_db()


}
