% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{test_cbioportal_db}
\alias{test_cbioportal_db}
\title{Test the Database Connection Anytime During your R Session}
\usage{
test_cbioportal_db()
}
\value{
No return value, called for side effects.
Will display an alert notifying if the user has successfully authenticated
to cBioPortal
}
\description{
Helps troubleshoot API issues during an R session
}
\examples{
\dontshow{if (!httr::http_error("www.cbioportal.org/api")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
set_cbioportal_db("public")
test_cbioportal_db()

\dontshow{\}) # examplesIf}
}
\author{
Karissa Whiting, Daniel D. Sjoberg
}
